better organization of cli commands code
