+++ src/cli/commands/start/types.ts
@@ -1,0 +1,6 @@
+import { BaseCliArgumentsOptions } from '../../types'
+
+export type StartCliArgumentsOptions = BaseCliArgumentsOptions & {
+  port?: string
+  host?: string
+}
