+++ src/comp-site/vue/index.ts
@@ -1,3 +1,4 @@
 import { createApp } from 'vue'
 import App from './App.vue'
+
 createApp(App).mount('#exh-root')
